Converted the number to int

+++ Collatz.py
@@ -51,7 +51,7 @@ def collatz_help (num, length):
 
     elif (num % 2 != 0):
         length += 1
-        num = ((3*num) + 1)/2
+        num = int(((3*num) + 1)/2)
         return collatz_helper (num)
 
     else:
